Release 2.0.8
* Update documentation links to point to GitHub

Replaced relative links with direct GitHub URLs in cf-bypass-cache.md, cf-smart-cache-workers.md, and cf-smart-cache.md for improved accessibility and clarity. Added a link to the cf-bypass-cache README and removed a redundant reference.

* Update package name and version

Changed the package name to 'cloudflare-smart-tools' and updated the version to 0.0.1 in package.json.

* Bump plugin version to 2.0.8

Updated the plugin version in the main PHP file from 2.0.7 to 2.0.8. No other changes included in this commit.